Move Cubic Feature Sampling to a new file.
