+++ torch_points_kernels/__init__.py
@@ -2,6 +2,7 @@ from .torchpoints import *
 from .knn import knn
 from .cluster import region_grow
 from .metrics import instance_iou
+from .cubic_feature_sampling import cubic_feature_sampling
 
 __all__ = [
     "ball_query",
@@ -13,4 +14,5 @@ __all__ = [
     "region_grow",
     "instance_iou",
     "chamfer_dist",
+    "cubic_feature_sampling",
 ]
